requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
groq>=0.11.0
python-dotenv>=1.0.1
python-multipart>=0.0.9
pydantic-settings>=2.4.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
